+++ src/controllers/ElementIndexesController.php
@@ -255,6 +255,20 @@ class ElementIndexesController extends BaseElementsController
         if ($success) {
             // Send a new set of elements
             $responseData = array_merge($responseData, $this->elementResponseData(true, true));
+
+            // Send updated badge counts
+            /** @var string|ElementInterface $elementType */
+            $elementType = $this->elementType;
+            $formatter = Craft::$app->getFormatter();
+            foreach ($elementType::sources($this->context) as $source) {
+                if (isset($source['key'])) {
+                    if (isset($source['badgeCount'])) {
+                        $responseData['badgeCounts'][$source['key']] = $formatter->asDecimal($source['badgeCount'], 0);
+                    } else {
+                        $responseData['badgeCounts'][$source['key']] = null;
+                    }
+                }
+            }
         }
 
         return $this->asJson($responseData);